Enabled warnAboutDeprecatedLifecycles flag by default
