+++ packages/react-dom/src/__tests__/ReactServerRenderingHydration-test.js
@@ -358,11 +358,21 @@ describe('ReactDOMServerHydration', () => {
     );
 
     const element = document.createElement('div');
-    element.innerHTML = ReactDOMServer.renderToString(markup);
+    expect(() => {
+      element.innerHTML = ReactDOMServer.renderToString(markup);
+    }).toLowPriorityWarnDev(
+      ['componentWillMount() is deprecated and will be removed'],
+      {withoutStack: true},
+    );
     expect(element.textContent).toBe('Hi');
 
-    expect(() => ReactDOM.hydrate(markup, element)).toWarnDev(
-      'Please update the following components to use componentDidMount instead: ComponentWithWarning',
+    expect(() => {
+      expect(() => ReactDOM.hydrate(markup, element)).toWarnDev(
+        'Please update the following components to use componentDidMount instead: ComponentWithWarning',
+      );
+    }).toLowPriorityWarnDev(
+      ['componentWillMount is deprecated and will be removed'],
+      {withoutStack: true},
     );
     expect(element.textContent).toBe('Hi');
   });
