+++ packages/react/src/__tests__/createReactClassIntegration-test.internal.js
@@ -9,119 +9,24 @@
 
 'use strict';
 
-let React;
-let ReactFeatureFlags;
-let createReactClass;
-
 describe('create-react-class-integration', () => {
-  beforeEach(() => {
-    jest.resetModules();
-
-    ReactFeatureFlags = require('shared/ReactFeatureFlags');
-    ReactFeatureFlags.warnAboutDeprecatedLifecycles = true;
-
-    React = require('react');
-    createReactClass = require('create-react-class/factory')(
-      React.Component,
-      React.isValidElement,
-      new React.Component().updater,
-    );
-  });
-
-  // TODO (RFC #6) Merge this back into createReactClassIntegration-test once
-  // the 'warnAboutDeprecatedLifecycles' feature flag has been removed.
-  it('isMounted works', () => {
-    const ReactDOM = require('react-dom');
-
-    const ops = [];
-    let instance;
-    const Component = createReactClass({
-      displayName: 'MyComponent',
-      mixins: [
-        {
-          UNSAFE_componentWillMount() {
-            this.log('mixin.componentWillMount');
-          },
-          componentDidMount() {
-            this.log('mixin.componentDidMount');
-          },
-          UNSAFE_componentWillUpdate() {
-            this.log('mixin.componentWillUpdate');
-          },
-          componentDidUpdate() {
-            this.log('mixin.componentDidUpdate');
-          },
-          componentWillUnmount() {
-            this.log('mixin.componentWillUnmount');
-          },
-        },
-      ],
-      log(name) {
-        ops.push(`${name}: ${this.isMounted()}`);
-      },
-      getInitialState() {
-        this.log('getInitialState');
-        return {};
-      },
-      UNSAFE_componentWillMount() {
-        this.log('componentWillMount');
-      },
-      componentDidMount() {
-        this.log('componentDidMount');
-      },
-      UNSAFE_componentWillUpdate() {
-        this.log('componentWillUpdate');
-      },
-      componentDidUpdate() {
-        this.log('componentDidUpdate');
-      },
-      componentWillUnmount() {
-        this.log('componentWillUnmount');
-      },
-      render() {
-        instance = this;
-        this.log('render');
-        return <div />;
-      },
-    });
-
-    const container = document.createElement('div');
-
-    expect(() => ReactDOM.render(<Component />, container)).toWarnDev(
-      'Warning: MyComponent: isMounted is deprecated. Instead, make sure to ' +
-        'clean up subscriptions and pending requests in componentWillUnmount ' +
-        'to prevent memory leaks.',
-      {withoutStack: true},
-    );
-
-    // Dedupe
-    ReactDOM.render(<Component />, container);
-
-    ReactDOM.unmountComponentAtNode(container);
-    instance.log('after unmount');
-    expect(ops).toEqual([
-      'getInitialState: false',
-      'mixin.componentWillMount: false',
-      'componentWillMount: false',
-      'render: false',
-      'mixin.componentDidMount: true',
-      'componentDidMount: true',
-      'mixin.componentWillUpdate: true',
-      'componentWillUpdate: true',
-      'render: true',
-      'mixin.componentDidUpdate: true',
-      'componentDidUpdate: true',
-      'mixin.componentWillUnmount: true',
-      'componentWillUnmount: true',
-      'after unmount: false',
-    ]);
-  });
-
   describe('ReactNative NativeMethodsMixin', () => {
+    let React;
     let ReactNative;
     let NativeMethodsMixin;
+    let createReactClass;
 
     beforeEach(() => {
+      jest.resetModules();
+
+      React = require('react');
+
+      createReactClass = require('create-react-class/factory')(
+        React.Component,
+        React.isValidElement,
+        new React.Component().updater,
+      );
+
       ReactNative = require('react-native-renderer');
       NativeMethodsMixin =
         ReactNative.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED
